+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Feature:** Pass auth data as props to Loading Component [#75](https://github.com/mjrussell/redux-auth-wrapper/issues/75)
 - **Breaking Change:** When rendering an "empty" component such as the default LoadingComponent and when the user is about
 to be redirect, return `null` so React will omit it instead of an empty `div`.
+- **Breaking Change:** Don't need to import React Native compatible from lib.
 
 ## [0.7.0](https://github.com/mjrussell/redux-auth-wrapper/compare/v0.6.0...v0.7.0)
 - **Bugfix:** Don't render subroutes as children in default LoadingComponent [#63](https://github.com/mjrussell/redux-auth-wrapper/pull/63)